requirements.txt
django==3.2.0
sqlparse==0.4.3
asgiref==3.6.0
psycopg2-binary~=2.9.3
python-dotenv==1.0.0
pytz==2023.3
backports.zoneinfo==0.2.1  
djangorestframework==3.14.0
